==> ukb_preprocessing/Preprocessing/05c Blood_Visual.Rmd <==
---
title: "Blood Visual"
output: html_document
date: "2024-04-30"
---


```{r}
source("../config.R") #Loads libraries, variables and global function
```



#### Define global and user-specific variables
1. Define your own paths and variables
2. R Detects your Microsoft / Apple/ Linux username (could result in conflicts with same name)
3. Assign your own variables to global variables
```{r}

# IOIs <-c("C240", "C241", "C248", "C249")
# IOI <- "C240"
# DOI <- "eCCa"
export_filename = paste("Blood", DOI)
date = format(Sys.time(), " %Y-%m-%d")
control <- toupper(risk_constellation)
dir.create(file.path(project_path, "/Blood Visuals"), showWarnings = FALSE)
path <- paste(project_path, "/Blood Visuals/", sep="")
date_out = format(Sys.time(), " %Y-%m-%d")        
file.out <- paste("Blood ", DOI, " vs ", control, sep="")  #file.out = Substring for the output file name.
fact <- "mets" 
output_format <- "SVG"  #either "SVG" or "PNG"

```




```{r}


setwd(sharepoint_ukb)
df_withdrawals <- read.delim("raw/withdrawals.txt")
covariates <- fread("extracted/covariates.csv") %>% check_and_remove_withdrawals(df_withdrawals)
covariate_index <- read_excel("meta/Master_Table_FVH.xlsx", sheet = "All covariates")   # import index data (Liste mit allen covariates, codes, units etc)
covariates <-  setnames(covariates, old = covariate_index$datafield1, new = covariate_index$Assessment, skip_absent=TRUE) #rename variables after index
covariates_subset <- covariates %>%
  select("eid", "Ethnicity", "BMI", "AGE", "SEX", "PC1", "PC2", "PC3", "PC4", "PC5")              #only covariates of interest
covariates_subset <- covariates_subset %>% filter(SEX == 0)
setwd(project_path)
load("data/dataframes/df_icd.RData") 
load("data/dataframes/df_par_5_years.RData")
df_icd <-df_icd %>% check_and_remove_withdrawals(df_withdrawals)
sapply(df_icd, class)


```


#### 2. Import and process Table Y (Information (Diagnosis, Factor etc. to be correlated)) (can also be taken from df_icd above)
Format has to be 

|--eid--|--DOI--|

after following markdown segment, with DOI containing 0 and 1 information for the diagnosis of interest
```{r}
#Import information that contains 0 - control group (matches) and 1 (phenotype/diagnosis of interest) from PPSM (1) or directly from UKB (Extrahiere_Diagnosen): Two columns, eid and DOI


# To switch between options, mark all the rows you want to run and press Strg + Shift + C


      # Option A1: from matching / big group / subgroup (Table Y Extract)
      setwd(project_path)
      df_y <- read.csv(paste(project_path, "/data/dataframes/df_y.csv", sep='')) %>% check_and_remove_withdrawals(df_withdrawals)
      df_doi<-select(df_y, c(eid, status)) %>% #Change HCC to diagnosis of your choice
        rename(c("DOI" = "status"))

      df_doi <- inner_join(df_doi, par_eids, by="eid") #Subset your desired subset (defined previously in Extract multiple diagnosis)
      rm(df_y)
      

      #Option A2: from matching/subgroup
      #setwd(home)
      #df_ppsm <- fread("Patient_tables/Cirrhosis_Subgroup_HCC.csv")
      #df_y <- read.csv(paste(project, "/data/dataframes/df_y.csv", sep=''))
      #df_doi<-select(df_y, c(eid, HCC)) %>% #Change HCC to diagnosis of your choice
        #rename(c("DOI" = "HCC"))
      
      #df_doi <- read.csv("HCC/Pat_with_HCC_C220_and_controls.csv")
      

      # #Option B: without matching, continuous variable -> gets transformed in high and low group
      # setwd("C:/Users/Jan/OneDrive/Dokumente/PostDoc/Patient_tables") 
      # df_doi <- fread("VitaminE.csv") 
      # 
      # median <- median(df_cont$VitaminE)
      # df_doi$DOI[df_cont$VitaminE > median] <- 1
      # df_doi$DOI[df_cont$VitaminE < median] <- 0
      # 
      # df_doi<-select(df_doi, c(eid, DOI))
      
      # #Option C: using only upper and lower quartile
      # setwd("C:/Users/Jan/OneDrive/Dokumente/PostDoc/Patient_tables") 
      # df_doi <- fread("VitaminE.csv")                                             #Changes these according to your DOI
      # df_doi$quartile <- ntile(df_doi$VitaminE, 4)                                #Changes these according to your DOI
      # df_doi$DOI[df_doi$quartile == 4] <- 1
      # df_doi$DOI[df_doi$quartile == 1] <- 0
      # df_doi$DOI[df_doi$quartile == 3] <- NA
      # df_doi$DOI[df_doi$quartile == 2] <- NA
      # df_doi <- df_doi[complete.cases(df_doi)]
      
      
      #Option D (Missing): Directly for continuous variable
      
      #Quality control
      print(paste("Amount of positive cases in imported df:", sum(df_doi$DOI, na.rm = TRUE)))

```


#### 2. Blood sample data for all 502411 patients df_blood 502411

```{r}
#setwd(drive)
#import blood sample data from baseline and change names to measured parameters
#df_blood <- fread("ukb52200.csv", select=c("eid", "30160-0.0","30220-0.0","30150-0.0","30210-0.0","30030-0.0","30020-0.0","30300-0.0","30290-0.0","30280-0.0","30120-0.0","30180-0.0","30050-0.0","30060-0.0","30040-0.0","30100-0.0","30260-0.0","30270-0.0","30130-0.0","30190-0.0","30140-0.0","30200-0.0","30170-0.0","30230-0.0","30080-0.0","30090-0.0","30110-0.0","30010-0.0","30070-0.0","30250-0.0","30240-0.0","30000-0.0","30620-0.0","30600-0.0","30610-0.0","30630-0.0","30640-0.0","30650-0.0","30710-0.0","30680-0.0","30690-0.0","30700-0.0","30720-0.0","30660-0.0","30730-0.0","30740-0.0","30750-0.0","30760-0.0","30770-0.0","30780-0.0","30790-0.0","30800-0.0","30810-0.0","30820-0.0","30830-0.0","30850-0.0","30840-0.0","30860-0.0","30870-0.0","30880-0.0","30670-0.0","30890-0.0"))
#setwd(sharepoint_ukb)
#write.csv(df_blood, file="extracted/UKB_Patients_blood.csv, row.names=FALSE)    

setwd(sharepoint_ukb)
df_blood <- fread("extracted/UKB_Patients_blood.csv") %>% check_and_remove_withdrawals(df_withdrawals)

# changing names to measured parameters
#Blood_Marker_Index <- read_excel("meta/Master_Table_JC.xlsx", sheet="Blood count and biochemistry")           #Import Index dataframe
Blood_Marker_Index <- read_excel(master_table, sheet="Blood count and biochemistry")    
Blood_Marker_Index$datafield <- paste0(Blood_Marker_Index$datafield, '-0.0')    #adapt syntax
blood_names <- Blood_Marker_Index$Description                                   #Create names vector
blood_names <- c("eid", blood_names)                                            #add eid in front of name vector
df_blood <- setNames(df_blood, blood_names)                                     #Change names from codes to actual measurements
df_blood <- setnames(df_blood, old=Blood_Marker_Index$Description, new = Blood_Marker_Index$Short_name, skip_absent=TRUE)



#write.xlsx(na_blood_col, file=paste(project, "/supplement/NA_count_df_blood.xlsx", sep=''))

summary(df_blood[,1:5])

save(df_blood, file=paste(project_path, "/data/dataframes/df_blood.RData", sep='')) # Save df_blood as .RData (preserves column classes)



#Sanity check
#sanity(df_blood)
```


#### 3b Merging tables with blood to table with actual outcome (DOI) and controls
```{r}
#perform inner join
      df_norm_met<-merge(df_blood, df_doi , by.x = "eid", by.y = "eid")
      print(paste("Amount of", DOI, "in Blood dataframe:", sum(df_norm_met$DOI, na.rm = TRUE)))
      print(paste("Amount of", control, "(controls) in Blood dataframe:", sum(df_norm_met$DOI == 0, na.rm = TRUE)))
      df_norm_met<-as.data.frame(df_norm_met)

#merge covariate data of interest and diagnosis to df for correcting on age, sex, bmi, diabetes etc.
      df_norm_met<-merge(df_norm_met, covariates_subset , by.x = "eid", by.y = "eid", all.x=TRUE)
      df_norm_met<-merge(df_norm_met, df_icd , by.x = "eid", by.y = "eid", all.x=TRUE)

```

#### 4. Linear model

```{r}
#We analyzed 249 parameters measured by NMR. To do so, we created a dataset that includes the 249 metabolites as well as age, BMI, sex, ethnicity and HCCstatus. 
summary_table <- function(x) {
  
  # Capture number of columns passed to the function
  num_vars <- ncol(x)
  
  # Pre-define lists that will be populated and then collapsed by rest of function
  models <- vector("list", length = num_vars)
  first_tables <- vector("list", length = num_vars)
  second_tables <- vector("list", length = num_vars)
  
  # Loop to create each row for the final table (Add other covariates here ( AGE + SEX + BMI))
  for (i in 1:num_vars) {
    
    #models[[i]] <- lm(x[[i]] ~ HCC + AGE + SEX + BMI + Other.and.unspecified.cirrhosis.of.liver, data = df_norm_met)                     #Controlling for Liver cirrhosis
     models[[i]] <- lm(x[[i]] ~ DOI + AGE + SEX + BMI, data = df_norm_met)
    first_tables[[i]] <- broom::tidy(models[[i]])
  } 
  
  # Combine the rows together into a final table
  final_table <- do.call("rbind", first_tables)
  
  return(final_table)
}
#In Klammer einfügen: [1.Zu untersuchender Parameter: Letzter Metabolit (= vorletzte Spalte) ], term anpassen zu Diagnosis of interest
final_table1 <- summary_table(df_norm_met[2:62])
final_table_DOI <- subset(final_table1, term=="DOI")
final_table_DOI$log <- -(log10(final_table_DOI$p.value))
final_table_DOI<-final_table_DOI[-c(1,2,3,4,5)]
```


#### 5b: Preparation for Circle Plot
```{r}
circleplot_index <- Blood_Marker_Index
table_circleplot <- subset(final_table1, term=="DOI")
table_circleplot$text1 <- circleplot_index$Short_name
table_circleplot<-as.data.frame(table_circleplot) %>%
  select(-term)
table_circleplot <- cbind(id_name_s = rownames(table_circleplot), table_circleplot)
table_circleplot <- rename(table_circleplot, c("Estimate" = "estimate"))
table_circleplot <- rename(table_circleplot, c("StdErr" = "std.error"))




write.xlsx(table_circleplot, file= paste(project_path,"/Blood Visuals/", export_filename,date, ".xlsx", sep = ""))

```



# Calculate optimal range of Y axis according to displayed estimates
```{r}
min_estimate <- min(table_circleplot$Estimate)
max_estimate <- max(table_circleplot$Estimate)
buffer <- (max_estimate - min_estimate) * 0.2
min_plot <- min_estimate - buffer
max_plot <- max_estimate + buffer



min_estimate
max_estimate
min_plot
max_plot

symlog <- function(x, C = 1) {
  sign(x) * log1p(abs(x) / C)
}

C_value <- 1  # Adjust C_value as needed based on your data's scale and domain

# Transform the estimates for plotting
min_plot_transformed <- symlog(min_plot, C_value)
max_plot_transformed <- symlog(max_plot, C_value)

# Plot limits
YLIM <- c(min_plot_transformed, max_plot_transformed)

# Calculate YCUTS based on transformed values
YCUTS <- seq(min_plot_transformed, max_plot_transformed, length.out = 5)
YCUTS.LABS <- round(exp(YCUTS), 2)

YCUTS
YLIM
```




```{r}

#################################################### 2. Import datasets ################################################

#Using the objects define above, we now call the datasource with SAS output to create datasub.

#We exponentiate Estimate to create RR.
#We then create RR_1<-RR if RR< 1 (i.e. those with negative associations), else RR_1<-1.
#We also then create RR_2<-RR if RR>1 (i.e. those with positive associations), else RR_2<-1.

datasub <<- table_circleplot

# 1.
datasub$RR <- exp(datasub$Estimate)

# 2.
datasub$RR_1 <- datasub$RR
datasub$RR_1[datasub$RR>1] <- 1

# 3.
datasub$RR_2 <- datasub$RR
datasub$RR_2[datasub$RR<1] <- 1


#########################################  3. Adjusting p-values #############################################

# 1. From SAS output now imported into datasub, estimate p-values from chisq statistics datasub$RawP.
# 2. Using the false discovery rate adjustment by Benjamini & Hochberg, p.adjust estimates adjusted p-values datasub$AdjP.
# 3. Then, adds flags for the metabolites with evidence against the null hypothesis bellow the fdr-adjusted “significance level”.
# 4. We then create new vectors for estimates that are significant (suffix = _s). 
#    NB, suffix _1 is used for estimates with negative associations and suffix _2 for estimates 
#    with positive associations. We will add colours later (red for positive and blue for negative, darker shade for those below the significance threshold).
# 5. If flagged as “non-significant” then newly created vectors are transformed into 1 (the value for the null hypothesis).
# 6. If flagged as “significant” then original vectors are transformed into 1 (the value for the null hypothesis).
# 7. Estimates the number of metabolites based on de dimension of the dataset, necessary later.

#1. (datasub$RawP <- pchisq(datasub$WaldChiSq, 1, lower.tail=FALSE)) -> Only necessary when input != p-value but WaldChiSquare (for categorical variables)

#2.
datasub$AdjP <- p.adjust(datasub$p.value, method = "fdr")

# 3.
datasub$Sig<-NA
datasub$Sig[datasub$AdjP< 0.05] <- 1
datasub$Sig[datasub$AdjP>= 0.05] <- 0

# 4. 
datasub$RR_1_s <- datasub$RR_1
datasub$RR_2_s <- datasub$RR_2

# 5. 
datasub$RR_1_s[datasub$Sig==0] <- 1
datasub$RR_2_s[datasub$Sig==0] <- 1

# 6.
datasub$RR_1[datasub$Sig==1] <- 1
datasub$RR_2[datasub$Sig==1] <- 1

# 7.
len.data <<-  as.numeric(dim(datasub)[1])




########################################  4. Plotting parameters ###########################################################

# In this section we input the parameters for the plotting areas and steps are taken to keep proportions. Importantly, the measures to keep proportionality could be substantially improved.

#### Y-axis ####

# 1. YLIM YCUTS and YCUTS.LABS define the Y-axis. Parameters here are defined manually but could be automated by extracting MIN and MAX 
#    and using the pretty function to define cuts and labels.
# 2. Alternatively, one could define labels as percentage instead of relative risks, if desired.
# 3. ylab 1:3 define the levels for labels around the circular plot that are relative and proportional to the MAX and MIN of the axis.
# 4. If estimate is off limits from YLIM then estimates are trimmed. Currently, the plot doesn’t flag this transformation, 
#    although it should be evident as the bar ends precisely at the limit of the axis and user should be aware as the axis limits are currently defined manually.

#Log_scale <- read_excel("C:/Users/Jan/OneDrive/Dokumente/PostDoc/Figures/Circle Plots/Log scale.xlsx")

# Alternativ: 0.6, 1.7 cuts= 0.6, 0.75, 1, 1.3, 1.7
YLIM <- c(log(0.01), log(100))
YCUTS <- c(log(0.01), log(0.1), log(1), log(10), log(100))
YCUTS.LABS <- as.character(exp(YCUTS))
YMAX <- exp(max(YLIM))
YMIN <- exp(min(YLIM))

#"zooming out"
#YLIM <- c(log(0.4), log(2.4))
#YCUTS <- c(log(0.4), log(0.6), log(1), log(1.67), log(2.4))
#YCUTS.LABS <- as.character(exp(YCUTS))
#YMAX <- exp(max(YLIM))
#YMIN <- exp(min(YLIM))

# 2.
#YCUTS.LABS <- c("-40%", "-20%", "0%", "30%", "60%")
#Change the parameters for varying y-position of labels (ylab3 = outer, 2 = medium (LDL etc), 1 = XXl, XL etc)
# 3.
ylab0 <- exp(log(YMAX)+log(YMAX)*0.1)
ylab1 <- exp(log(YMAX)+log(YMAX)*0.2)
ylab2 <- exp(log(YMAX)+log(YMAX)*0.7)
ylab3 <- exp(log(YMAX)+log(YMAX)*1.5)
ylab3b <- exp(log(YMAX)+log(YMAX)*1.3)

# 4.
ADJ <- 0.01

datasub$RR_1[datasub$RR_1<=YMIN] <- YMIN+YMIN*ADJ
datasub$RR_2[datasub$RR_2>=YMAX] <- YMAX-YMAX*ADJ

datasub$RR_1_s[datasub$RR_1_s<=YMIN] <- YMIN+YMIN*ADJ
datasub$RR_2_s[datasub$RR_2_s>=YMAX] <- YMAX-YMAX*ADJ

datasub$Estimate[datasub$Estimate<=log(YMIN)] <- log(YMIN)+log(YMIN)*ADJ
datasub$Estimate[datasub$Estimate>=log(YMAX)] <- log(YMAX)-log(YMAX)*ADJ


#### X-axis ####

# IMPORTANT the x-axis is defined by the number of metabolic biomarkers. This number is currently 139 derived from id_name_s. 
# All the labels are mapped around this number, and in this especific order. 
# If the user decides a different array of biomarkers is needed (i.e. only include lipids, or by lipid types instead of by lipoprotein sizes), 
# then this change can only currently be implemented in SAS and the mapping for labels should also be changed manually.

XLIM <- c(min(as.numeric(datasub$id_name_s)), max(as.numeric(datasub$id_name_s)))

#### Labels ####

# labs1 Contains the lipoprotein subclass size acronyms. This is repeated 7 times, once per each measurement of interest (i.e. lipoprotein particle number, cholesterol, free cholesterol, esterified cholesterol, triglycerides, phospholipids, and total lipids).
# labs4 Vector with additional labels for the rest of biomarkers besides lipids within lipoproteins.
# CEX states a vector to use for sizing. If user changes CEX (with upper case), then all those functions using CEX will be proportionally re-sized.
# NOTE if the user changes the array defining id_name_s, then this section should be changed accordingly.

labs1 <- Blood_Marker_Index$Short_name



#### Graphic device ####
# We decided to use the png graphic device, but others such as pdf or tiff do the trick as well.
# 1. The line we would use produces a filename that includes the file.out substring defined above as well as GROUP and the date.
# 2. For this document, we have named the output file "foo.png".
setwd(path)

output_format <- "SVG"
# 1.
if (output_format == "PNG") {
  png(paste(path, file.out, date, ".png", sep=""), height=6000, width=6000, bg = "transparent")
  CEX <- 9
  lwd <- 1.5
} else if (output_format == "SVG") {
  svg(paste(path, file.out, date, ".svg", sep=""), width = 40, height = 40, pointsize = 12, bg="transparent")
  CEX <- 2.3
  lwd <- 1
}



 #png(filename = "test2.png", height=6000, width=6000, bg = "white")




#### Margins ####
# The outer margins OMA are quite large (53 spaces, in the 4 margins), as we need space to place our labels.

par(xpd = NA, oma = rep(60,4))






#####################################################  Circos function ###########################################################

# This represents the core of the script, although most of the job is done above. It uses the circlize package but, as you will see, most of the basic R plot functions are preserved and only slightly changed.

# Importantly, this plot uses only very limitedly the applications of the circlize package. Some of the approaches I have had to make the plot are probably clumsy or redundant.

# I highly recommend to have a quick look into the documentation (https://jokergoo.github.io/circlize_book/book/). It is simpler than it looks and relatively easy to work with.


##### Parameters #####

# Circlize transforms a “Cartesian plane” with x and y axis into a circle of y radius and x circumference. 
# Basically, a traditional rectangular plot is twisted into a donut. The donut is called sector. Sectors can be split in several tracks. 
# You can add additional sectors or donuts in ever more central levels.

# Our example only has 1 sector with 1 track.
# track.height determines the proportion of the radius of the circle the track (where we are going to plot) is going to use. 
# The circle used by circlize always has a radius of 1, so a height of 0.1 means 10% of the circle radius.
# gap.degree determines the space between the end of the track and the start of the track.
# start.degree determines the place to start the track at in degrees (count starts at the West).

circos.par("track.height" = 0.6, 
           cell.padding = c(0, 0, 0, 0),
           gap.degree = 90,
           start.degree = 90,
           unit.circle.segments=50000,
           points.overflow.warning=FALSE)

##### Initialize the circle #####

# circos.initialize is the core function that determines the basic parameters. I am still not entirely sure how it works. 
# However, a character object in the factors option (in this example fact) does the trick and this becomes the name of our sector.
# xlim is defined by the length of the id_name_s column, as noted above.

# 1.
circos.initialize(factors = fact, xlim = c(0,len.data))

# 2.
circos.track(factors = fact, ylim = YLIM, bg.border = NA)

##### Draw shades for metabolic subgroups ##### ############################ Eventually to be changed due to addition of 4 markers

# To highlight specific regions use circlize() to calculate the positions in the polar coordinate. Always keep in mind that x-axis in the cell are always clock wise.
# The highlight region to be calculated by circlize()needs coordinates in x and y, a sector.index (in this case "mets"), and a track.index (in this case 1).
# NOTE: In this example, the coordinates were imputed manually and correspond to the array defined by id_names_s. If changed, this section must also be changed to preserve meaningful highlight regions.
# Unless the user wants to change the order of the biomarkers, this section needs no further details explained.

pos1 = circlize(c(0.5, 4.5), c(min(YLIM), max(YLIM)), sector.index = "mets", track.index = 1)
draw.sector(pos1[1, "theta"], pos1[2, "theta"], pos1[1, "rou"], pos1[2, "rou"], clock.wise = TRUE, col = "#CCCCCC56", border = NA) 

pos2 = circlize(c(8.5, 11.5), c(min(YLIM), max(YLIM)), sector.index = "mets", track.index = 1)
draw.sector(pos2[1, "theta"], pos2[2, "theta"], pos2[1, "rou"], pos2[2, "rou"], clock.wise = TRUE, col = "#CCCCCC56", border = NA)

pos3 = circlize(c(14.5, 20.5), c(min(YLIM), max(YLIM)), sector.index = "mets", track.index = 1)
draw.sector(pos3[1, "theta"], pos3[2, "theta"], pos3[1, "rou"], pos3[2, "rou"], clock.wise = TRUE, col = "#CCCCCC56", border = NA)

pos4 = circlize(c(24.5, 31.5), c(min(YLIM), max(YLIM)), sector.index = "mets", track.index = 1)
draw.sector(pos4[1, "theta"], pos4[2, "theta"], pos4[1, "rou"], pos4[2, "rou"], clock.wise = TRUE, col = "#CCCCCC56", border = NA)

pos5 = circlize(c(37.5, 42.5), c(min(YLIM), max(YLIM)), sector.index = "mets", track.index = 1)
draw.sector(pos5[1, "theta"], pos5[2, "theta"], pos5[1, "rou"], pos5[2, "rou"], clock.wise = TRUE, col = "#CCCCCC56", border = NA)

pos6 = circlize(c(46.5, 50.5), c(min(YLIM), max(YLIM)), sector.index = "mets", track.index = 1)
draw.sector(pos6[1, "theta"], pos6[2, "theta"], pos6[1, "rou"], pos6[2, "rou"], clock.wise = TRUE, col = "#CCCCCC56", border = NA)


pos7 = circlize(c(53.5, 57.5), c(min(YLIM), max(YLIM)), sector.index = "mets", track.index = 1)
draw.sector(pos7[1, "theta"], pos7[2, "theta"], pos7[1, "rou"], pos7[2, "rou"], clock.wise = TRUE, col = "#CCCCCC56", border = NA)




##### Plotting region #####

# 1. Using circos.track, we select track 1, using factors defined in object fact, and the YLIM defined above.
# 2. We use circos.segmets exactly as segments would be used to create:
  # I. Start and end of plot lines.
  # II. Outer and inner lines.
  # III.Lines at null hypothesis and other cuts.

# 1.
circos.track(track.index = 1, bg.border = "white", factors = fact, ylim = YLIM, panel.fun = function(x,y){
  
  # i)
  # Start 
  circos.segments(x0=min(XLIM)-0.5, y0=max(YLIM), x1=min(XLIM)-0.5, y1=min(YLIM), col = "black", lwd=lwd)
  # End 
  circos.segments(x0=max(XLIM)+0.5, y0=max(YLIM), x1=max(XLIM)+0.5, y1=min(YLIM), col = "black", lwd=lwd)
  
  # ii)
  # Outer
  circos.segments(x0=min(XLIM)-.75, y0=max(YLIM), x1=max(XLIM)+0.5, y1=max(YLIM), col = "black", lwd=lwd)
  # Inner
  circos.segments(x0=min(XLIM)-.75, y0=min(YLIM), x1=max(XLIM)+0.5, y1=min(YLIM), col = "black", lwd=lwd)
  
  # iii)
  # Lines at YCUTS
  # Null Hypothesis
  circos.segments(x0=min(XLIM)-.75, y0=0, x1=max(XLIM)+0.5, y1=0, col = "black", lwd=lwd)
  circos.segments(x0=min(XLIM)-.75, y0=(YCUTS[2]), x1=max(XLIM)+0.5, y1=(YCUTS[2]), col = "gray75", lwd=lwd)
  circos.segments(x0=min(XLIM)-.75, y0=(YCUTS[4]), x1=max(XLIM)+0.5, y1=(YCUTS[4]), col = "gray75", lwd=lwd)
  
  
  
  
#### Draw bars with estimates ####
  
  # circos.rect draws a rectangle of xleft, xright, ytop, and ybottom dimentions.
  
  # Each bar is defined in the x axis by its position withing is_name_s. Width is defined by simply substracting or adding 0.35 to the coordinates in xleft and xright, respectively.
  
  # Each bar of the 4 types of bars are defined in the y axis by the value in one of the four RR vectors created above, based on the following:
    
    # 1. Positive and “significant”, in dark red (i.e. RR_2_s).
    # 2. Positive and not “significant”, in light red (i.e. RR_2).
    # 3. Negative and “significant”, in dark blue (i.e. RR_1_s).
    # 4. Negative and “non-significant”, in light blue (i.e. RR_1).
  
  # Colours are defined in hex with the last 2 digits defining transparency.

  # Bars
  # 1.
  circos.rect(xleft=(as.numeric(datasub$id_name_s)-.35), xright=(as.numeric(datasub$id_name_s)+.35), ytop=log(as.numeric(datasub$RR_2_s)), ybottom = log(1), col = "#CC0000CC" , lwd=lwd)
  # 2.
  circos.rect(xleft=(as.numeric(datasub$id_name_s)-.35), xright=(as.numeric(datasub$id_name_s)+.35), ytop=log(as.numeric(datasub$RR_2)), ybottom = log(1), col = "#CC000040" , lwd=lwd)
  # 3.
  circos.rect(xleft=(as.numeric(datasub$id_name_s)-.35), xright=(as.numeric(datasub$id_name_s)+.35),ytop=log(as.numeric(datasub$RR_1_s)), ybottom = log(1), col = "#0066CCCC" , lwd=lwd)
  # 4.
  circos.rect(xleft=(as.numeric(datasub$id_name_s)-.35), xright=(as.numeric(datasub$id_name_s)+.35), ytop=log(as.numeric(datasub$RR_1)), ybottom = log(1), col = "#0066CC40" , lwd=lwd)

  
  
  # NOTE: The following chunk must be plotted after the bars, so the graphic device can draw the confidence intervals on top.
circos.rect(
    xleft = as.numeric(datasub$id_name_s),
    xright = as.numeric(datasub$id_name_s),
    ytop = pmin(as.numeric(datasub$Estimate) + 1.95 * as.numeric(datasub$StdErr), YLIM[2]),
    ybottom = pmax(as.numeric(datasub$Estimate) - 1.95 * as.numeric(datasub$StdErr), YLIM[1]),
    col = "#262626",
    lwd = lwd
)

  circos.points(x=as.numeric(datasub$id_name_s[datasub$Sig == 1]), y=log(ylab0), pch = 42, col = "#000000", cex = CEX*1.0) #### Display significance stars ####
  
  #### Draw y-axis ####
  # Similar to basic R plotting axis options.
  
  circos.yaxis(at=YCUTS, labels = YCUTS.LABS, labels.cex = CEX*1, tick = FALSE, col = "white")
  
#### Draw labels ####
  # We use the function circos.text to paste the labels at the margins of the plot (remember we left a lot of space at the margins when defining the plot par above).
  
  # The option facing defines how to paste the labels. The package has several options that make text look nicely, including niceFacing, which makes text flip so it can be read easily.
  
  # The positions for labels are, unfortunately, very inefficiently defined manually.
  
  # All the .shift objects were used to manually adjust the labels. These work now, but maybe play with them to see how the labels move.
  
  
  VLDL.shift <- 2 
  LDL.shift <- 1.5
  HDL.shift <- 1.5
  
  ylab3_size <- 0.9
  ylab2_size <- 0.9
  ylab1_size <- 0.8
  

  
  #Inner Labels
  circos.text(x=c(1:62)+0.25,   y = log(ylab1), labels = labs1, facing = "clockwise", niceFacing = TRUE, cex = CEX*ylab1_size, adj = 0, font = 1)
  
}

)


#### Title ####
 
if (length(control) ==1){
  text(0, 0, paste(control, "\n±\n", DOI, sep=""), cex = CEX*1.1, font=2)
} else{
  text(0, 0, paste("Blood Samples","\n","Patients at Risk", "\n±\n", DOI, sep=""), cex = CEX*1.1, font=2)
}
dev.off()  # Close the plotting device with dev.off().
circos.clear()   # Use circos.clear to reset the circular layout parameters.



############################################################### 6. Output ###################################################

#png()

## 2
```

```
